refactor proposal content markdown section from Vote page into ProposalContent component
