+++ packages/nouns-webapp/src/pages/Vote/Vote.module.css
@@ -41,33 +41,6 @@
   font-family: 'Londrina Solid';
 }
 
-.markdown {
-  font-family: 'PT Root UI';
-  font-size: 1.1rem;
-}
-
-.markdown h1 {
-  font-size: 1.7rem;
-  margin-top: 1rem;
-  font-weight: bold;
-}
-
-.markdown h2 {
-  font-size: 1.5rem;
-  margin-top: 1rem;
-  font-weight: bold;
-}
-
-.markdown h3 {
-  font-weight: bold;
-  font-size: 1.3rem;
-}
-
-.markdown img {
-  max-width: 100%;
-  height: auto;
-}
-
 .voterIneligibleAlert {
   margin: 1rem 0 0 0;
 }
